Add somes control in back end #1

+++ pAPI/Controllers/AddressController.cs
@@ -1,9 +1,12 @@
 ﻿using System;
+using Application.Repositories;
 using Application.Services.Address;
 using Application.Services.Address.Dto;
 using Application.Services.AddressUser;
 using Application.Services.AddressUser.Dto;
+using Application.Services.Users;
 using Microsoft.AspNetCore.Mvc;
+using NUnit.Framework.Constraints;
 
 namespace pAPI.Controllers
 {
@@ -16,41 +19,61 @@ namespace pAPI.Controllers
     {
         private readonly IAddressService _addressService;
         private readonly IAddressUserService _addressUserService;
+        private readonly IUserRepository _userRepository;
+        private readonly IAddressRepository _addressRepository;
 
-        public AddressController(IAddressService addressService, IAddressUserService addressUserService)
+        public AddressController(IAddressService addressService, IAddressUserService addressUserService, IUserRepository userRepository, IAddressRepository addressRepository)
         {
             _addressService= addressService;
             _addressUserService = addressUserService;
+            _userRepository = userRepository;
+            _addressRepository = addressRepository;
         }
 
         //ActionResult renvoie un code http et entre <> c'est les données qui vont être renvoyées
         [HttpGet]
         public ActionResult<OutputDtoQueryAddress> QueryAddress()
         {
+
+            var query = _addressService.Query();
+
+            if (query == null)
+                return BadRequest(new {message = "Aucune données n'a été trouvées."});
             //Renvoie les données avec un code 200 -> Tout s'est bien passé
-            return Ok(_addressService.Query());
+            return Ok(query);
         }
 
         [HttpGet]
         [Route("{id}")]
         public ActionResult<OutputDtoGetByIdAddress> GetByIdAddress(int id)
         {
+            if (id < 0)
+            {
+                return BadRequest(new {message = "L'id n'est pas conforme."});
+            }
             var inputDtoGetByIdAddress = new InputDtoGetByIdAddress
             {
                 id = id
             };
             
             OutputDtoGetByIdAddress address = _addressService.GetById(inputDtoGetByIdAddress);
-            return _addressService!= null ? (ActionResult<OutputDtoGetByIdAddress>) Ok(address) : NotFound();
+            return _addressService!= null ? (ActionResult<OutputDtoGetByIdAddress>) Ok(address) : BadRequest(new {message = "Aucune adresse ne correspond à l'id envoyé."});
         }
 
+        //TODO Vérifier l'intégralité des données envoyées pour le post
         //[FromBody] le user qu'on enverra, résidera dans le corps de la requête
         [HttpPost]
         public ActionResult<OutputDtoAddAddress> CreateAddress([FromBody]InputDtoAddAddress inputDtoAddAddress)
         {
-            return Ok(_addressService.Create(inputDtoAddAddress));
+            var address = _addressService.Create(inputDtoAddAddress);
+            if (address == null)
+            {
+                BadRequest(new {message = "L'adresse n'a pas été créée."});
+            }
+            return Ok(address);
         }
         
+        //TODO Vérifier l'intégralité des données envoyées pour le post
         //Post pour l'enregistrement d'une voiture et d'une adresse d'un utilisateur
         [HttpPost]
         [Route("addressCar")]
@@ -62,7 +85,14 @@ namespace pAPI.Controllers
             };
             if (_addressUserService.GetByIdUserAddressUser(inputDtoGetByIdUserAddressUser) != null)
                 return BadRequest(new {message = "Vous êtes déjà enregistré en tant que covoitureur"});
-            return Ok(_addressService.CreateAddressAndCarByid(inputDtoAddAddressAndCar));
+
+            var addresscar = _addressService.CreateAddressAndCarByid(inputDtoAddAddressAndCar);
+
+            if (addresscar == null)
+            {
+                BadRequest(new {message = "L'adresse et la voiture n'ont pas été créée"});
+            }
+            return Ok(addresscar);
             
         }
 
@@ -70,19 +100,31 @@ namespace pAPI.Controllers
         [Route("{id}")]
         public ActionResult DeleteByIdAddress(int id)
         {
+            if (id < 0)
+            {
+                return BadRequest(new {message = "L'id n'est pas conforme."});
+            }
+            
             var inputDtoDeleteByIdAddress = new InputDtoDeleteByIdAddress()
             {
                 id = id
             };
             
+            if (_addressRepository.GetById(inputDtoDeleteByIdAddress.id) == null)
+            {
+                return BadRequest(new
+                    {message = "L'adresse n'existe pas."});
+            }
+
             if (_addressService.DeleteById(inputDtoDeleteByIdAddress))
             {
                 return Ok();
             }
 
-            return NotFound();
+            return BadRequest(new {message = "L'adresse n'a pas été supprimée"});
         }
 
+        //TODO Vérifier l'intégralité des données envoyées pour l'update
         [HttpPut]
         [Route("{id}")]
         public ActionResult UpdateAddress(int id,[FromBody]InputDtoUpdateAddress inputDtoUpdateAddress)
@@ -92,37 +134,78 @@ namespace pAPI.Controllers
                 return Ok();
             }
 
-            return NotFound();
+            return BadRequest(new {message = "L'adresse n'a pas été modifiée"});
         }
 
         [HttpGet]
         [Route("{idAddress}/users")]
         public ActionResult<OutputDtoGetByIdAddressAddressUser> GetByIdAdressUser(int idAddress)
         {
+            if (idAddress < 0)
+            {
+                return BadRequest(new {message = "L'id n'est pas conforme."});
+            }
+            
             var inputDtoGetByIdAddressAddressUser = new InputDtoGetByIdAddressAdressUser
             {
                 IdAddress = idAddress
             };
+            
+            if (_addressRepository.GetById(inputDtoGetByIdAddressAddressUser.IdAddress) == null)
+            {
+                return BadRequest(new
+                    {message = "L'adresse n'existe pas."});
+            }
 
-            return Ok(_addressUserService.GetByIdAddressAddressUser(inputDtoGetByIdAddressAddressUser));
+            var user = _addressUserService.GetByIdAddressAddressUser(inputDtoGetByIdAddressAddressUser);
+
+            if (user == null)
+            {
+                BadRequest(new {message = "Aucun n'utilisateur ne correspond à l'adresse donnée"});
+            }
+
+            return Ok(user);
         }
         
         [HttpGet]
         [Route("{idUser}/address")]
         public ActionResult<OutputDtoGetByIdUserAddressUser> GetByIdUser(int idUser)
         {
+            if (idUser < 0)
+            {
+                return BadRequest(new {message = "L'id n'est pas conforme."});
+            }
+            
             var inputDtoGetByIdUserAddressUser = new InputDtoGetByIdUserAddressUser
             {
                 IdUser = idUser
             };
+            
+            if (_userRepository.GetById(inputDtoGetByIdUserAddressUser.IdUser) == null)
+            {
+                return BadRequest(new
+                    {message = "L'utilisateur n'existe pas."});
+            }
 
-            return Ok(_addressUserService.GetByIdUserAddressUser(inputDtoGetByIdUserAddressUser));
+            var address = _addressUserService.GetByIdUserAddressUser(inputDtoGetByIdUserAddressUser);
+
+            if (address == null)
+            {
+                BadRequest(new {message = "Aucune n'adresse ne correspond à l'utilisateur donné"});
+            }
+
+            return Ok(address);
         }
 
         [HttpPost]
         [Route("{idUser}/{idAddress}/users")]
         public ActionResult<OutputDtoCreateAddressUser> CreateAddressUser(int idUser, int idAddress)
         {
+            if (idUser < 0 || idAddress < 0)
+            {
+                return BadRequest(new {message = "L'id n'est pas conforme."});
+            }
+            
             var inputDtoIdUserCreateAddressUser = new InputDtoIdUserCreateAddressUser
             {
                 IdUser = idUser
@@ -131,8 +214,27 @@ namespace pAPI.Controllers
             {
                 IdAddress = idAddress
             };
-            Console.WriteLine(idUser);
-            return Ok(_addressUserService.CreateAddressUser(inputDtoIdUserCreateAddressUser,inputDtoIdAddressCreateAddressUser));
+
+            if (_userRepository.GetById(inputDtoIdUserCreateAddressUser.IdUser) == null)
+            {
+                return BadRequest(new
+                    {message = "L'utilisateur n'existe pas."});
+            }
+            
+            if (_addressRepository.GetById(inputDtoIdAddressCreateAddressUser.IdAddress) == null)
+            {
+                return BadRequest(new
+                    {message = "L'adresse n'existe pas."});
+            }
+
+            var addressuser = _addressUserService.CreateAddressUser(inputDtoIdUserCreateAddressUser,
+                inputDtoIdAddressCreateAddressUser);
+
+            if (addressuser == null)
+                return BadRequest(new
+                    {message = "La correspondance adresse user n'a pas été créée dans la base de donnée"});
+            
+            return Ok(addressuser);
         }
     }
 }
